pyproject.toml
[tool.poetry]
name = "macrolens"
version = "0.1.0"
description = ""
authors = ["Jason Hoelscher-Obermaier <jason.hoelscherobermaier@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = {extras = ["openai"], version = "^0.0.232"}
transformers = "^4.30.2"
tqdm = "^4.65.0"
gymnasium = "^0.28.1"
pettingzoo = "^1.23.1"
python-dotenv = "^1.0.0"
streamlit = "^1.24.1"
hydra-core = "^1.3.2"
gitpython = "^3.1.32"
gptcache = "^0.1.35"
setuptools = "^68.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-vcr = "^1.0.2"
jupyterlab = "^4.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
